modify for car package review fix

+++ app/Http/Controllers/Admin/BonusController.php
@@ -11,6 +11,7 @@ use App\Models\User;
 use App\Models\Bonus;
 use App\Models\BonusCompleteList;
 use App\Http\Lib\Helper;
+use Carbon\Carbon;
 use DB;
 
 class BonusController extends Controller
@@ -114,8 +115,16 @@ class BonusController extends Controller
     public function destroy (Request $request) 
     {
         $bonus = Bonus::findOrFail($request->id);
-        $bonus->delete();
-        return redirect()->route('bonus.index')->with('message', 'Bonus deleted successfully');
+        
+        $bonus_exist_in_complete = BonusCompleteList::where('bonus_id', $request->id)->first();
+        
+        if (!$bonus_exist_in_complete) {
+            $bonus->delete();
+            return redirect()->route('bonus.index')->with('message', 'Bonus deleted successfully');
+        } else {
+            return redirect()->route('bonus.index')->with('error_message', 'Sorry, this bonus has transaction');
+        }
+        
     } 
     
     /**
@@ -123,41 +132,48 @@ class BonusController extends Controller
     */
     public function capable (Request $request) 
     {  
-        if ($request->type == 0) { //0 mean = user
+        $current_date_time = Carbon::now()->toDateTimeString(); 
+        $bonus = Bonus::find($request->bonus_id);
+        if ($current_date_time <= $bonus->offer_finishing_time) {
+            if ($request->type == 0) { //0 mean = user
         
-            $records = DB::table('ride_list')
-                        ->join('users','ride_list.user_id','users.id')
-                        ->select('users.id','users.name','users.phone', DB::raw("COUNT(ride_list.id) as total_completed"))
-                        ->where('ride_list.start_time', '>=', $request->start)
-                        ->where('ride_list.start_time', '<=', $request->end)
-                        ->where('ride_list.status', 4)
-                        ->groupBy('ride_list.id','users.id','users.name','users.phone')
-                        ->get();
-                        
-            $records = $records->where('total_completed', $request->completed);
+                $records = DB::table('ride_list')
+                            ->join('users','ride_list.user_id','users.id')
+                            ->select('users.id','users.name','users.phone', DB::raw("COUNT(ride_list.id) as total_completed"))
+                            ->where('ride_list.start_time', '>=', $request->start)
+                            ->where('ride_list.start_time', '<=', $request->end)
+                            ->where('ride_list.status', 4)
+                            ->groupBy('ride_list.id','users.id','users.name','users.phone')
+                            ->get();
+                            
+                $records = $records->where('total_completed', $request->completed);
+                
+            } else { 
+                
+                $records = DB::table('ride_biting')
+                            ->leftjoin('ride_list','ride_biting.ride_id','ride_list.id')
+                            ->leftjoin('owners','ride_biting.owner_id','owners.id')
+                            ->select('owners.id','owners.name','owners.phone',DB::raw("COUNT(ride_list.id) as total_completed"))
+                            ->where('ride_biting.status', 1)
+                            ->where('ride_list.accepted_ride_bitting_id', '!=', null)
+                            ->where('ride_list.start_time', '>=', $request->start)
+                            ->where('ride_list.start_time', '<=', $request->end)
+                            ->where('ride_list.status', 4)
+                            ->groupBy('ride_list.id','owners.id','owners.name','owners.phone')
+                            ->get();
+                            
+                $records = $records->where('total_completed', $request->completed);
+            }
             
-        } else { 
-            
-            $records = DB::table('ride_biting')
-                        ->leftjoin('ride_list','ride_biting.ride_id','ride_list.id')
-                        ->leftjoin('owners','ride_biting.owner_id','owners.id')
-                        ->select('owners.id','owners.name','owners.phone',DB::raw("COUNT(ride_list.id) as total_completed"))
-                        ->where('ride_biting.status', 1)
-                        ->where('ride_list.accepted_ride_bitting_id', '!=', null)
-                        ->where('ride_list.start_time', '>=', $request->start)
-                        ->where('ride_list.start_time', '<=', $request->end)
-                        ->where('ride_list.status', 4)
-                        ->groupBy('ride_list.id','owners.id','owners.name','owners.phone')
-                        ->get();
-                        
-            $records = $records->where('total_completed', $request->completed);
+            $type = $request->type;
+            $required_completed = $request->completed;
+            $bonus_id = $request->bonus_id;
+         
+            return view('quicarbd.admin.bonus.capable', compact('records', 'type', 'required_completed','bonus_id'));
+        } else {
+            return redirect()->back()->with('error_message','Bonus already expired');
         }
         
-        $type = $request->type;
-        $required_completed = $request->completed;
-        $bonus_id = $request->bonus_id;
-     
-        return view('quicarbd.admin.bonus.capable', compact('records', 'type', 'required_completed','bonus_id'));
     }
     
     /**
